src/main/java/com/example/order_management_system/service/ProductService.java
package com.example.order_management_system.service;

import com.example.order_management_system.model.Product;
import com.example.order_management_system.repository.ProductRepository;
import org.springframework.stereotype.Service;

@Service
public class ProductService {

    private final ProductRepository productRepository;

    public ProductService(ProductRepository productRepository){
        super();
        this.productRepository = productRepository;
    }

    public Product saveProduct(Product product){
        return productRepository.save(product);
    }

}
